requirements.txt
streamlit>=1.32.0
openai>=1.14.0
python-dotenv>=1.0.0
Pillow>=10.2.0